Moves two tests from miscellaneous to verifier.
